+++ apps/rx-store-website/src/components/Apidocs.js
@@ -1,12 +1,17 @@
 import React from 'react';
 import Layout from '@theme/Layout';
 
-export default ({ __content, __menu }) => {
+export default ({ __content, __menu, __title }) => {
   return (
     <Layout title="API docs">
       <div class="container">
         <div class="row">
-          <div class="col col--8">
+          <div class="col col--8" style={{ paddingTop: 15, paddingBottom: 15 }}>
+            <div dangerouslySetInnerHTML={{ __html: __title }} />
+          </div>
+        </div>
+        <div class="row">
+          <div class="col col--8" style={{ paddingTop: 15, paddingBottom: 15 }}>
             <div dangerouslySetInnerHTML={{ __html: __content }} />
           </div>
           <div class="col col--4">