+++ libs/core/src/lib/sinks.ts
@@ -1,8 +1,8 @@
-import { StoreValue } from '..';
+import { StoreValue } from './store-value';
 import { debug } from 'debug';
 import { tap } from 'rxjs/operators';
-import { Observable, MonoTypeOperatorFunction, Observer } from 'rxjs';
-import { StoreEvent, StoreEventType } from './store-arg';
+import { Observable, MonoTypeOperatorFunction } from 'rxjs';
+import { StoreEventType, StoreObserver } from './store-event';
 
 /**
  * Sinks are a write-only interface into the subjects
@@ -15,17 +15,22 @@ import { StoreEvent, StoreEventType } from './store-arg';
  *
  * The next method from the subject is then wrapped in an operator
  * such that it can be consumed by referencing it from a pipeline
+ *
+ * This is a mapped type in Typescript, meaning it has the same
+ * type as the {@link StoreValue} except the values of the object
+ * described by the type are mapped to `next()` methods, which essentially
+ * "seals them off" and makes them write only.
  */
 export type Sinks<T extends StoreValue> = {
   [P in keyof T]: () => MonoTypeOperatorFunction<Parameters<T[P]['next']>[0]>;
 };
 
 /**
- * The createSinks method takes a debug key (string), store value
+ * The createSinks method takes a debug key (string), {@link StoreValue}
  * (object containing subjects), and returns an object with the same
- * keys as the original store value, but instead of the subjects
- * themselves the values are a "next" method, which is curried with
- * the debug key & forwards the events to the underlying subjects.
+ * keys as the original {@link StoreValue} passed in to the method,
+ * but instead of the subjects themselves the values are a "next" method, which is
+ * curried with the debug key & forwards the events to the underlying subjects.
  *
  * Each effect should get its own sinks object, with its own debug key.
  *
@@ -33,11 +38,13 @@ export type Sinks<T extends StoreValue> = {
  * @param storeValue The Rx Store value, object containing subjects
  * @returns An object matching the shape of the original storeValue, but with "next" methods
  * instead of the subjects themselves.
+ *
+ * @internal
  */
 export const createSinks = <T extends StoreValue>(
   effectName: string,
   storeValue: T,
-  observer?: Observer<StoreEvent>
+  observer?: StoreObserver
 ): Sinks<T> => {
   return Object.keys(storeValue).reduce(
     (acc, subjectName) => ({