+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ### Features
 
 - Makes spawnEffect generic so users can type projected values [#72](https://github.com/rx-store/rx-store/pull/72)
+- Adds typedoc generated API reference to docs [#81](https://github.com/rx-store/rx-store/pull/81)
+
+### Code Refactoring
+
+- Adds TSdoc comments to code, adds `StoreObserver`, `Manager` types, splits up files [#82](https://github.com/rx-store/rx-store/pull/82),
 
 # [1.1.2](https://github.com/rx-store/rx-store/compare/1.1.0...1.1.2) (09-03-2020)